requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-multipart
fitz
pytesseract
Pillow
numpy
opencv-python
sentence-transformers
chromadb
python-docx
google-generativeai
python-dotenv
weasyprint
ollama
aiohttp
